+++ packages/updating-element/CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Lifecycle callbacks added for connected, disconnected, update, and updated. To use, simply add callback functions to the appropriate callbacks set: `connectedCallbacks`, `disconnectedCallbacks`, `updateCallbacks`, or `updatedCallbacks`.
+
 - UpdatingElement moved from `lit-element` package to `updating-element` package.
 
 ### Fixed